--- src/main/java/io/jenkins/plugins/analysis/core/QualityGateEnforcer.java
@@ -1,19 +1,0 @@
-package io.jenkins.plugins.analysis.core;
-
-import hudson.model.Result;
-
-/**
- * Enforces the defined quality gates.
- *
- * @author Ullrich Hafner
- */
-public class QualityGateEnforcer {
-    public Result evaluate(final StaticAnalysisRun run, final QualityGate qualityGate) {
-        if (qualityGate.hasFailureThreshold()) {
-            if (run.getTotalHighPrioritySize() >= qualityGate.getFailureThreshold()) {
-                return Result.FAILURE;
-            }
-        }
-        return Result.SUCCESS;
-    }
-}
