+++ src/main/java/io/jenkins/plugins/analysis/core/history/NullBuildHistory.java
@@ -1,26 +1,43 @@
 package io.jenkins.plugins.analysis.core.history;
 
-import java.util.Optional;
+import java.util.Calendar;
+import java.util.GregorianCalendar;
 
-import io.jenkins.plugins.analysis.core.steps.AnalysisResult;
+import io.jenkins.plugins.analysis.core.steps.BuildResult;
+
+import hudson.model.Run;
 
 /**
  * Empty build history.
  *
  * @author Ulli Hafner
  */
-// FIXME: still required?
 public class NullBuildHistory extends BuildHistory {
     /**
      * Creates a new instance of {@link NullBuildHistory}.
      */
     public NullBuildHistory() {
-        super(null, null);
+        super((Run<?, ?>)null, null);
     }
 
     @Override
-    public Optional<AnalysisResult> getPreviousResult() {
-        return Optional.empty();
+    public Calendar getTimestamp() {
+        return new GregorianCalendar();
+    }
+
+    @Override
+    public boolean hasPrevious() {
+        return false;
+    }
+
+    @Override
+    public boolean isEmpty() {
+        return true;
+    }
+
+    @Override
+    public BuildResult getPrevious() {
+        return null;
     }
 }
