+++ src/main/java/io/jenkins/plugins/analysis/core/history/NullReferenceProvider.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.analysis.core;
+package io.jenkins.plugins.analysis.core.history;
 
 import javax.annotation.CheckForNull;
