+++ src/main/java/io/jenkins/plugins/analysis/core/history/ReferenceFinder.java
@@ -1,81 +1,87 @@
 package io.jenkins.plugins.analysis.core.history;
 
-import java.util.Optional;
+import javax.annotation.CheckForNull;
 
-import edu.hm.hafner.analysis.Issues;
-import io.jenkins.plugins.analysis.core.steps.ResultAction;
+import io.jenkins.plugins.analysis.core.steps.PipelineResultAction;
 
-import hudson.model.Result;
 import hudson.model.Run;
+import hudson.plugins.analysis.util.model.AnnotationContainer;
+import hudson.plugins.analysis.util.model.DefaultAnnotationContainer;
 
 /**
- * Finds a previous result of an analysis run for the same software artifact. Selection of the previous result is
- * delegated to a provided strategy.
+ * FIXME: write comment.
  *
+ * @since 2.0
  * @author Ullrich Hafner
- * @see ResultSelector
  */
 public abstract class ReferenceFinder extends BuildHistory implements ReferenceProvider {
-    /**
-     * Creates a {@link ReferenceProvider} instance based on the specified properties.
-     *
-     * @param run
-     *         the run to use as baseline when searching for a reference
-     * @param selector
-     *         selects the type of the result (to get a result for the same type of static analysis)
-     * @param ignoreAnalysisResult
-     *         if {@code true} then the result of the previous analysis run is ignored when searching for the reference,
-     *         otherwise the result of the static analysis reference must be {@link Result#SUCCESS}.
-     * @param overallResultMustBeSuccess
-     *         if  {@code true} then only runs with an overall result of {@link Result#SUCCESS} are considered as a
-     *         reference, otherwise every run that contains results of the same static analysis configuration is
-     *         considered
-     */
-    @SuppressWarnings("BooleanParameter")
     public static ReferenceProvider create(final Run<?, ?> run, final ResultSelector selector,
-            final boolean ignoreAnalysisResult, final boolean overallResultMustBeSuccess) {
-        if (ignoreAnalysisResult) {
-            return new PreviousRunReference(run, selector, overallResultMustBeSuccess);
+            final boolean usePreviousBuildAsReference, final boolean useStableBuildAsReference) {
+        if (usePreviousBuildAsReference) {
+            return new PreviousBuildReference(run, selector, useStableBuildAsReference);
         }
         else {
-            return new StablePluginReference(run, selector, overallResultMustBeSuccess);
+            return new StablePluginReference(run, selector, useStableBuildAsReference);
         }
     }
 
     /**
-     * Creates a new instance of {@link ReferenceFinder}.
+     * Creates a new instance of {@link BuildHistory}.
      *
      * @param baseline
-     *         the run to start the history from
+     *            the build to start the history from
      * @param selector
-     *         selects the type of the result (to get a result for the same type of static analysis)
+     *            selects the associated action from a build
      */
-    protected ReferenceFinder(final Run<?, ?> baseline, final ResultSelector selector) {
+    public ReferenceFinder(final Run<?, ?> baseline, final ResultSelector selector) {
         super(baseline, selector);
     }
 
     /**
      * Returns the action of the reference build.
      *
-     * @return the action of the reference build, or {@code null} if no such build exists
+     * @return the action of the reference build, or {@code null} if no
+     *         such build exists
      */
-    protected abstract Optional<ResultAction> getReferenceAction();
+    protected abstract PipelineResultAction getReferenceAction();
+
+    /**
+     * Returns whether a reference build result exists.
+     *
+     * @return <code>true</code> if a reference build result exists.
+     */
+    private boolean hasReferenceAction() {
+        return getReferenceAction() != null;
+    }
+
+    @Override @CheckForNull
+    public Run<?, ?> getReference() {
+        PipelineResultAction action = getReferenceAction();
+        if (action != null) {
+            Run<?, ?> build = action.getRun();
+            if (hasValidResult(build)) {
+                return build;
+            }
+        }
+        return null;
+    }
 
     @Override
-    public int getNumber() {
-        return getReferenceAction().map(pipelineResultAction -> pipelineResultAction.getRun().getNumber())
-                .orElse(NO_REFERENCE_FOUND);
+    public boolean hasReference() {
+        return getReference() != null;
     }
 
     /**
-     * Returns the issues of the reference build.
+     * Returns the annotations of the reference build.
      *
-     * @return the issues of the reference build
+     * @return the annotations of the reference build
      */
     @Override
-    public Issues getIssues() {
-        return getReferenceAction()
-                .map(resultAction -> resultAction.getResult().getIssues())
-                .orElseGet(Issues::new);
+    public AnnotationContainer getIssues() {
+        PipelineResultAction action = getReferenceAction();
+        if (action != null) {
+            return action.getResult().getContainer();
+        }
+        return new DefaultAnnotationContainer();
     }
 }
