+++ src/main/java/io/jenkins/plugins/analysis/core/history/StablePluginReference.java
@@ -1,43 +1,38 @@
 package io.jenkins.plugins.analysis.core.history;
 
-import java.util.Optional;
+import io.jenkins.plugins.analysis.core.steps.PipelineResultAction;
 
-import io.jenkins.plugins.analysis.core.steps.ResultAction;
-
-import hudson.model.Result;
 import hudson.model.Run;
 
 /**
- * Find the last available analysis run with a result {@link Result#SUCCESS} for the specified analysis configuration.
- * You can also specify if the overall result of the run has a result of {@link Result#SUCCESS}.
+ * FIXME: write comment.
  *
  * @author Ullrich Hafner
  */
 public class StablePluginReference extends ReferenceFinder {
-    private final boolean overallResultMustBeSuccess;
+    private boolean mustBeStable;
 
     /**
-     * Creates a new instance of {@link StablePluginReference}.
+     * Creates a new instance of {@link PreviousBuildReference}.
      *
-     * @param baseline
-     *         the build to start the history from
-     * @param selector
-     *         type of the action that contains the build results
-     * @param overallResultMustBeSuccess
-     *         if  {@code true} then only runs with an overall result of {@link Result#SUCCESS} are considered as a
-     *         reference, otherwise every run that contains results of the same static analysis configuration is
-     *         considered
+     * @param baseline     the build to start the history from
+     * @param selector     type of the action that contains the build results
+     * @param mustBeStable builds must be of overall status stable
      */
-    @SuppressWarnings("BooleanParameter")
-    public StablePluginReference(final Run<?, ?> baseline, final ResultSelector selector,
-            final boolean overallResultMustBeSuccess) {
+    public StablePluginReference(final Run<?, ?> baseline, final ResultSelector selector, final boolean mustBeStable) {
         super(baseline, selector);
-
-        this.overallResultMustBeSuccess = overallResultMustBeSuccess;
+        this.mustBeStable = mustBeStable;
     }
 
     @Override
-    protected Optional<ResultAction> getReferenceAction() {
-        return getPreviousAction(false, overallResultMustBeSuccess);
+    protected PipelineResultAction getReferenceAction() {
+        PipelineResultAction action = getAction(true, mustBeStable);
+        if (action == null) {
+            return getPreviousAction(); // fallback, use action of previous build regardless of result
+        }
+        else {
+            return action;
+        }
     }
+
 }
