+++ src/main/java/io/jenkins/plugins/analysis/core/quality/HealthDescriptor.java
@@ -4,8 +4,10 @@ import java.io.Serializable;
 
 import org.apache.commons.lang.StringUtils;
 
-import edu.hm.hafner.analysis.Priority;
 import static hudson.plugins.analysis.util.ThresholdValidator.*;
+
+import hudson.plugins.analysis.util.ThresholdValidator;
+import hudson.plugins.analysis.util.model.Priority;
 
 /**
  * A health descriptor defines the parameters to create the build health.
@@ -31,22 +33,22 @@ public class HealthDescriptor implements Serializable {
      * Returns the healthy threshold, i.e. when health is reported as 100%.
      *
      * @return the 100% healthiness
-     * @throws IllegalArgumentException
-     *         if the healthy string can't be converted to an integer value greater or equal zero
+     * @throws IllegalArgumentException if the healthy string can't be converted to an integer value greater or equal
+     *                                  zero
      */
     public int getHealthy() {
-        return convert(healthy);
+        return ThresholdValidator.convert(healthy);
     }
 
     /**
      * Returns the unhealthy threshold, i.e. when health is reported as 0%.
      *
      * @return the 0% unhealthiness
-     * @throws IllegalArgumentException
-     *         if the unhealthy string can't be converted to an integer value greater or equal zero
+     * @throws IllegalArgumentException if the unhealthy string can't be converted to an integer value greater or equal
+     *                                  zero
      */
     public int getUnHealthy() {
-        return convert(unHealthy);
+        return ThresholdValidator.convert(unHealthy);
     }
 
     /**
@@ -62,10 +64,10 @@ public class HealthDescriptor implements Serializable {
     /**
      * Determines whether a health report should be created.
      *
-     * @return {@code true} if a health report should be created
+     * @return <code>true</code> if a health report should be created
      */
     public boolean isEnabled() {
-        if (isValid(healthy) && isValid(unHealthy)) {
+        if (ThresholdValidator.isValid(healthy) && ThresholdValidator.isValid(unHealthy)) {
             int healthyNumber = convert(healthy);
             int unHealthyNumber = convert(unHealthy);
