+++ src/main/java/io/jenkins/plugins/analysis/core/quality/StaticAnalysisRun.java
@@ -1,9 +1,5 @@
 package io.jenkins.plugins.analysis.core.quality;
 
-import java.util.Map;
-
-import edu.hm.hafner.analysis.Priority;
-
 /**
  * Describes the results of a static analysis run.
  *
@@ -16,16 +12,6 @@ public interface StaticAnalysisRun {
      * @return total number of issues
      */
     int getTotalSize();
-
-    /**
-     * Returns the total number of issues in this analysis run, that have the specified {@link Priority}.
-     *
-     * @param priority
-     *         the priority of the issues to match
-     *
-     * @return total number of issues
-     */
-    int getTotalSize(Priority priority);
 
     /**
      * Returns the total number of high priority issues in this analysis run.
@@ -49,51 +35,30 @@ public interface StaticAnalysisRun {
     int getTotalLowPrioritySize();
 
     /**
-     * Returns the number of new issues in this analysis run.
+     * Returns the new number of issues in this analysis run.
      *
-     * @return number of new issues
+     * @return new number of issues
      */
     int getNewSize();
 
     /**
-     * Returns the number of new high priority issues in this analysis run.
+     * Returns the new number of high priority issues in this analysis run.
      *
-     * @return number of new high priority issues
+     * @return new number of high priority issues
      */
     int getNewHighPrioritySize();
 
     /**
-     * Returns the number of new normal priority issues in this analysis run.
+     * Returns the new number of normal priority issues in this analysis run.
      *
-     * @return number of new normal priority issues
+     * @return new number of normal priority issues
      */
     int getNewNormalPrioritySize();
 
     /**
-     * Returns the number of new low priority issues in this analysis run.
+     * Returns the new number of low priority issues in this analysis run.
      *
-     * @return number of new low priority of issues
+     * @return new number of low priority of issues
      */
     int getNewLowPrioritySize();
-
-    /**
-     * Returns the number of fixed issues in this analysis run.
-     *
-     * @return number of fixed issues
-     */
-    int getFixedSize();
-
-    /**
-     * Returns the number of issues in this analysis run, mapped by their origin.
-     *
-     * @return number of issues per origin
-     */
-    Map<String, Integer> getSizePerOrigin();
-
-    /**
-     * Returns the associated build that this run was part of.
-     *
-     * @return the associated build
-     */
-    AnalysisBuild getBuild();
 }
